fix: 修复SystemMonitorService中CPU使用率计算的整数溢出问题
- 将getCPUUsage()中的累加变量从UInt32改为UInt64
- 修复长时间运行系统(41天+)在多核处理器上的崩溃问题
- CPU状态计数器是累积值,UInt32在长时间运行后会溢出
- 崩溃异常: EXC_BREAKPOINT (SIGTRAP) in getCPUUsage()

+++ desktop/macos/share-my-status-client/Services/SystemMonitorService.swift
@@ -163,17 +163,17 @@ actor SystemMonitorService: PollingMonitoringService {
             vm_deallocate(mach_task_self_, vm_address_t(bitPattern: cpuInfo), vm_size_t(numCpuInfo))
         }
         
-        var totalUser: UInt32 = 0
-        var totalSystem: UInt32 = 0
-        var totalIdle: UInt32 = 0
-        var totalNice: UInt32 = 0
+        var totalUser: UInt64 = 0
+        var totalSystem: UInt64 = 0
+        var totalIdle: UInt64 = 0
+        var totalNice: UInt64 = 0
         
         for i in 0..<Int(numCpus) {
             let cpuLoadInfo = cpuInfo.advanced(by: Int(CPU_STATE_MAX) * i)
-            totalUser += UInt32(cpuLoadInfo[Int(CPU_STATE_USER)])
-            totalSystem += UInt32(cpuLoadInfo[Int(CPU_STATE_SYSTEM)])
-            totalIdle += UInt32(cpuLoadInfo[Int(CPU_STATE_IDLE)])
-            totalNice += UInt32(cpuLoadInfo[Int(CPU_STATE_NICE)])
+            totalUser += UInt64(cpuLoadInfo[Int(CPU_STATE_USER)])
+            totalSystem += UInt64(cpuLoadInfo[Int(CPU_STATE_SYSTEM)])
+            totalIdle += UInt64(cpuLoadInfo[Int(CPU_STATE_IDLE)])
+            totalNice += UInt64(cpuLoadInfo[Int(CPU_STATE_NICE)])
         }
         
         let totalTicks = totalUser + totalSystem + totalIdle + totalNice
